requirements.txt
torch 
numpy
pandas
openai-whisper
torchaudio
tqdm
ipywidgets
transformers[torch]
ffmpeg
